+++ packages/soda-heroui/rslib.config.ts
@@ -1,5 +1,3 @@
 import { getRslibConfig } from "../../utils/getRslibConfig"
 
-export default getRslibConfig({
-    reactCompiler: false,
-})
+export default getRslibConfig()