+++ utils/getRslibConfig.ts
@@ -1,14 +1,9 @@
 import { readFile } from "fs/promises"
 import { RsbuildPlugins } from "@rsbuild/core"
-import { pluginBabel } from "@rsbuild/plugin-babel"
 import { pluginReact } from "@rsbuild/plugin-react"
 import { defineConfig } from "@rslib/core"
 
-export interface GetRslibConfigParams {
-    reactCompiler?: boolean
-}
-
-export async function getRslibConfig({ reactCompiler = true }: GetRslibConfigParams = {}) {
+export async function getRslibConfig() {
     const text = await readFile("./package.json", "utf-8")
     const packageJson = JSON.parse(text)
     const isReact = (packageJson.dependencies?.react ?? packageJson.devDependencies?.react ?? packageJson.peerDependencies?.react) !== undefined
@@ -17,16 +12,6 @@ export async function getRslibConfig({ reactCompiler = true }: GetRslibConfigPar
 
     if (isReact) {
         plugins.push(pluginReact())
-        if (reactCompiler) {
-            plugins.push(
-                pluginBabel({
-                    babelLoaderOptions(config) {
-                        config.plugins ??= []
-                        config.plugins.unshift("babel-plugin-react-compiler")
-                    },
-                }),
-            )
-        }
     }
 
     return defineConfig({