+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-hooks
 
+## 6.15.0
+
+### Minor Changes
+
+- 移除 react compiler
+
 ## 6.14.6
 
 ### Patch Changes