+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,16 @@
 # soda-react-router
 
+## 1.2.0
+
+### Minor Changes
+
+- 移除 react compiler
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.15.0
+
 ## 1.1.48
 
 ### Patch Changes